backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.10.1
python-dotenv==1.0.1
gunicorn==23.0.0
Werkzeug==3.1.3
python-socketio==5.10.0
python-engineio==4.8.0
Flask-SocketIO==5.3.6
google-auth==2.30.0
requests==2.32.3
